feat: implement CSS variable caching
