+++ builder/www/builder_assets/color_scheme_variables.css
@@ -1,0 +1,6 @@
+[data-prefers-color-scheme="dark"] {
+{%- for key, value in dark_mode_css_variables.items() %}{{ key }}: {{ value }};{%- endfor %}}
+
+{%- if dark_mode_css_variables -%}[data-prefers-color-scheme="light"] {
+{%- for key, value in css_variables.items() %}{{ key }}: {{ value }};{%- endfor %}}
+{%- endif -%}
